% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visprov.R
\name{prov.visualize}
\alias{prov.visualize}
\title{prov.visualize}
\usage{
prov.visualize(r.script.path = NULL, tool = NULL, ...)
}
\arguments{
\item{r.script.path}{The path to an R script.  This script will be 
executed with provenance captured by the specified tool.  If r.script.path
is NULL, the last ddg captured will be displayed.}

\item{tool}{If an R script is passed in, this is the tool that will be used
to collect provenance.  Currently, the known choices are "provR" and 
"RDataTracker", which can be given as "rdt".  If no tool name is passed in,
provR will be used if it is loaded.  If provR is not loaded and RDataTracker
is loaded, RDataTracker will be used.  If neither has been loaded, it then checks
to see if either is installed.  If provR is installed, it will be use.  If 
provR is not installed but RDataTracker is installed, RDataTracker will be
used.  If neither is installed, an error is reported.}

\item{...}{If r.script.path is set, these parameters will passed to prov.run to 
collect the provenance.}
}
\description{
Display a provenance graph visually.
}
\examples{
\dontrun{prov.visualize ()}
\dontrun{prov.visualize ("script.R")}
\dontrun{prov.visualize ("script.R", tool = "provR")}
}
