% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visprov.R
\name{prov.visualize}
\alias{prov.visualize}
\title{prov.visualize}
\usage{
prov.visualize(r.script.path = NULL, tool = NULL, ...)
}
\arguments{
\item{r.script.path}{The path to an R script.  This script will be 
executed with provenance captured by the specified tool.  If r.script.path
is NULL, the last ddg captured will be displayed.}

\item{tool}{If an R script is passed in, this is the tool that will be used
to collect provenance.  Currently, the known choices are "rdt" and 
"rdtLite".  If no tool name is passed in,
rdtLite will be used if it is loaded.  If rdtLite is not loaded and rdt
is loaded, rdt will be used.  If neither has been loaded, it then checks
to see if either is installed.  If rdtLite is installed, it will be used.  If 
rdtLite is not installed but rdt is installed, rdt will be
used.  If neither is installed, an error is reported.}

\item{...}{If r.script.path is set, these parameters will be passed to prov.run to 
control how provenance is collected.  
See rdt's prov.run function
or rdtLites's prov.run function for details.}
}
\description{
Display a provenance graph visually.
}
\examples{
\dontrun{prov.visualize ()}
\dontrun{prov.visualize ("script.R")}
\dontrun{prov.visualize ("script.R", tool = "rdtLite")}
}
