% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visprov.R
\name{prov.visualize}
\alias{prov.visualize}
\title{prov.visualize}
\usage{
prov.visualize(r.script.path = NULL, tool = "provR")
}
\arguments{
\item{r.script.path}{The path to an R script.  This script will be 
executed with provenance captured by the specified tool.  If r.script.path
is NULL, the last ddg captured will be displayed.}

\item{tool}{If an R script is passed in, this is the tool that will be used
to collect provenance.  Currently, the known choices are "provR" and 
"RDataTracker", which can be given as "rdt".  The string is case-insensitive.}
}
\description{
Display a provenance graph visually.
}
\examples{
\dontrun{prov.visualize ()}
\dontrun{prov.visualize ("script.R")}
\dontrun{prov.visualize ("script.R", tool = "RDataTracker")}
}
